+++ src/util/uuid.hpp
@@ -10,7 +10,7 @@ class UUID
 public:
   static uint64_t generateUuid()
   {
-    uint64_t uuid = std::chrono::system_clock::now().time_since_epoch().count();
+    uint64_t uuid = static_cast<uint64_t>(std::chrono::system_clock::now().time_since_epoch().count());
     std::random_device rd;
     uuid = (uuid << 16) | rd();
     return uuid;
